# docs/swagger.yaml
components:
  schemas:
...
    api.commentResponse:
      properties:
        author_avatar:
          description: 作者头像地址
          type: string
        author_username:
          description: 作者用户名
          type: string
        content:
          description: 评论内容
          type: string
        created_at:
          description: 创建时间
          type: string
        id:
          description: 评论 ID
          type: string
        like_count:
          description: 点赞数
          minimum: 0
          type: integer
        parent_id:
          description: 父评论 ID,顶级评论为 nil
          type: string
        post_id:
          description: 帖子 ID
          type: string
        user_id:
          description: 用户 ID
          type: string
        viewer_liked:
          description: 当前用户是否已点赞
          type: boolean
      required:
      - id
      - like_count
      - post_id
      - user_id
      - viewer_liked
      type: object
    api.createCommentRequest:
      properties:
        content:
          description: 评论内容
          maxLength: 1000
          type: string
        parent_id:
          description: 父评论 ID,如果是顶级评论则为 nil
          type: string
        post_id:
          description: 帖子唯一标识 ID
          type: string
      required:
      - content
      - post_id
      type: object
...
    api.deleteCommentResponse:
      properties:
        comment_count:
          minimum: 0
          type: integer
        id:
          type: string
        post_id:
          type: string
      required:
      - comment_count
      - id
      - post_id
      type: object
...
    api.pageResponse-api_commentResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/api.commentResponse'
          type: array
          uniqueItems: false
        page:
          minimum: 1
          type: integer
        page_size:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      type: object
...
    render.Response-api_commentResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.commentResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_deleteCommentResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.deleteCommentResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_pageResponse-api_commentResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.pageResponse-api_commentResponse'
        msg:
          type: string
      type: object
...
    render.errorResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
...
info:
  description: 小蓝书后端接口文档
  title: 小蓝书 API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /comments:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.createCommentRequest'
              description: 评论内容
              summary: body
        description: 评论内容
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_commentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - comments
  /comments/{comment_id}:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: comment_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_deleteCommentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - comments
...
  /posts/{post_id}/comments:
    get:
      parameters:
      - description: 帖子 ID
        in: path
        name: post_id
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_pageResponse-api_commentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取帖子评论列表
      tags:
      - comments
...
servers:
- description: 本地开发环境
  url: http://localhost:8080/api/v1